app/api/agents/math_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-agent"
version = "0.1.0"
description = "An AI agent for solving mathematical problems and verifying LaTeX expressions"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Math Agent Team" }
]
dependencies = [
    "langchain-openai>=0.0.2",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.15",
    "httpx>=0.24.1",
    "pydantic>=2.0.0",
    "click>=8.1.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "pytest>=7.3.1",
] 